+++ example/main.tf
@@ -19,6 +19,7 @@ module "minio" {
   secret_key                      = data.vault_generic_secret.minio_secrets.data.secret_key
   data_dir                        = "/local/data"
   container_environment_variables = ["SOME_VAR_N1=some-value"]
+  use_host_volume                 = true
 
   # minio client
   mc_service_name                 = "mc"